Write affected modules to file, to only build those modules
